# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movies API",
        "description": "The Movies API allows you to manage a collection of movies with CRUD operations. It supports creating, retrieving, updating, and deleting movie records, including uploading and managing cover images.",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/movies": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Retrieve a list of movies",
                "description": "Display a listing of the resource.",
                "operationId": "a7859dc5cd106bbeae5492e22df893ec",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token for authorization",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer your-access-token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with a list of movies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Inception"
                                                    },
                                                    "release_date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2010-07-16"
                                                    },
                                                    "genre": {
                                                        "type": "string",
                                                        "example": "Science Fiction"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error response if there is an exception",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error message here"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Movies"
                ],
                "summary": "Create a new movie",
                "operationId": "a303e8fe5382b02a193fa994ebef739f",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token for authorization",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer your-access-token"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "image",
                                    "published_date"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Inception"
                                    },
                                    "image": {
                                        "description": "Movie image file",
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "published_date": {
                                        "type": "string",
                                        "format": "integer",
                                        "example": "2010-07-16"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Movie created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Movie has been created"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Inception"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "/storage/images/1629208800.jpg"
                                                },
                                                "published_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2010-07-16"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-08-28T12:34:56Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-08-28T12:34:56Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error response if there is a validation error or exception",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error message here"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/movies/{id}": {
            "put": {
                "tags": [
                    "Movies"
                ],
                "summary": "Update an existing movie",
                "description": "Update the specified resource in storage.",
                "operationId": "aea46ecba0a2c9a86915ba24cbc23184",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the movie to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token for authorization",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer your-access-token"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "published_date"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Inception"
                                    },
                                    "image": {
                                        "description": "Movie image file",
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "published_date": {
                                        "type": "string",
                                        "format": "integer",
                                        "example": "2010-07-16"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Movie updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Movie has been updated"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Inception"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "/storage/images/1629208800.jpg"
                                                },
                                                "published_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2010-07-16"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-08-28T12:34:56Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-08-28T12:34:56Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error response if there is a validation error or exception",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error message here"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}